requirements.txt
python-magic==0.4.27
Pillow==9.0.1